WIDEN-251 Handle consistency on DAM CDN. (#89)
* WIDEN-251 Handle consistency on DAM CDN.

When using transcoding, the DAM CDN can take up to an hour to become
consistent. We delay cron sync by an hour of an asset's last edit to
account for this. If transcoding is not used, the delay does not occur
and the asset is downloaded directly.

* WIDEN-251 use -1 as before cutoff after last sync time

* WIDEN-251 Fix services.yml

* WIDEN-251 Ignore LowerCamelName to match asset properties object

* WIDEN-251 Fix search query; implement preflight check on cut off time vs last sync time

* WIDEN-251 use HTTP client for downloading over file_get_contents

* WIDEN-251 prevent duplicate downloads due to strict checks

Timestamp returned from `::get` is a string, `strtotime` returns int. To
be compatible between 7.3-8.1 we use `!=` over `!==`.